+++ src/core/reportError.ts
@@ -1,4 +1,4 @@
-import getOptions from './getOptions';
+import getOptions from '../utils/getOptions';
 
 export default function reportError(message: string) {
     if (getOptions().onError) {
