+++ src/validation/validateFile.ts
@@ -1,14 +1,14 @@
-import NormalizedPath from './types/NormalizedPath';
-import normalizePath from './normalizePath';
-import TypeScriptProgram from './TypeScriptProgram';
-import validateImportIsAccessible from './validateImportIsAccessible';
+import NormalizedPath from '../types/NormalizedPath';
+import normalizePath from '../utils/normalizePath';
+import TypeScriptProgram from '../core/TypeScriptProgram';
+import validateExportRules from './validateExportRules';
 
 export default function validateFile(filePath: NormalizedPath, tsProgram: TypeScriptProgram) {
     const importedFiles = tsProgram.getImportsForFile(filePath);
     importedFiles.forEach(importInfo => {
         const resolvedFileName = tsProgram.resolveImportFromFile(importInfo.fileName, filePath);
         if (resolvedFileName) {
-            validateImportIsAccessible(filePath, normalizePath(resolvedFileName));
+            validateExportRules(filePath, normalizePath(resolvedFileName));
         }
     });
 }
